include for size_t

+++ ft_bzero.c
@@ -1,9 +1,9 @@
-#include <stdlib.h>
+#include <stddef.h>
 
 void    ft_bzero(void *s, size_t n)
 {
     char    *ptr;
-    int     i;
+    size_t     i;
 
     ptr = s;
     while (i < n)
